refactor shell and other lib
