+++ vantara/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [lib]
-path = "src/systemd/src/lib.rs"
+path = "src/lib/lib.rs"
 
 [dependencies]
 glob = "0.3"
