+++ vantara/src/commands/shell/Cargo.toml
@@ -12,4 +12,3 @@ dirs = "6.0.0"
 nix = "0.30.1"
 procfs = "0.17.0"
 vantara = { path = "../../../" }
-systemd = { path = "../../systemd" }
