+++ vantara/src/systemd/Cargo.toml
@@ -3,8 +3,12 @@ name = "systemd"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+path = "src/lib.rs"
+
 [dependencies]
 chrono = "0.4"
 chrono-tz="0.10.4"
 glob = "0.3"
 nix = { version = "0.30", features = ["process", "signal"] }
+vantara = { path = "../../" }
